+++ src/main/java/com/gym/login/IntroduceController.java
@@ -2,8 +2,6 @@ package com.gym.login;
 
 
 import com.gym.avatar.AvatarService;
-import com.gym.avatar.avatar.AvatarStep;
-import com.gym.avatar.avatar.MyAvatar;
 import com.gym.avatar.avatar.MyAvatarRepository;
 import com.gym.config.exception.BaseException;
 import com.gym.config.exception.BaseResponse;
@@ -11,11 +9,9 @@ import com.gym.config.exception.BaseResponseStatus;
 import com.gym.user.User;
 import com.gym.user.UserRepository;
 import com.gym.user.UserService;
-import com.gym.utils.JwtService;
+import com.gym.login.jwt.JwtService;
 import com.gym.utils.UtilService;
 import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestParam;
