+++ src/main/java/com/gym/login/jwt/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.gym.login;
+package com.gym.login.jwt;
 
 
 import com.gym.login.dto.JwtResponseDTO;
@@ -16,8 +16,8 @@ import java.util.Date;
 @Slf4j
 @Component
 public class JwtProvider {
-    private static final long REFRESH_TOKEN_EXPIRE_TIME = 14 * 24 * 60 * 60 * 1000L;
-    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1 * 6 * 60 * 60 * 1000L;
+    private static final long REFRESH_TOKEN_EXPIRE_TIME = 14 * 24 * 60 * 60 * 1000L; //refreshToken 유효기간 14일
+    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1 * 6 * 60 * 60 * 1000L; //accessToken 유효기간 6시간
     private static final String BEARER_TYPE = "Bearer";
 
     private Key key = Keys.hmacShaKeyFor(Decoders.BASE64URL.decode(Secret.JWT_SECRET_KEY));
