+++ src/main/java/com/gym/post/comment/CommentController.java
@@ -4,11 +4,9 @@ import com.gym.config.exception.BaseException;
 import com.gym.config.exception.BaseResponse;
 import com.gym.post.comment.dto.GetCommentsRes;
 import com.gym.post.comment.dto.PostCommentReq;
-import com.gym.utils.JwtService;
+import com.gym.login.jwt.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.repository.query.Param;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
