+++ src/main/java/com/gym/record/RecordService.java
@@ -10,7 +10,7 @@ import com.gym.record.photo.RecordPhotoService;
 import com.gym.tag.TagService;
 import com.gym.user.User;
 import com.gym.user.UserService;
-import com.gym.utils.JwtService;
+import com.gym.login.jwt.JwtService;
 import com.gym.utils.S3Service;
 import com.gym.utils.UtilService;
 import com.gym.utils.dto.getS3Res;
@@ -19,12 +19,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
