+++ src/main/java/com/gym/tag/TagService.java
@@ -6,7 +6,7 @@ import com.gym.config.exception.BaseException;
 import com.gym.config.exception.BaseResponseStatus;
 import com.gym.tag.dto.TagGetRecentRes;
 import com.gym.user.User;
-import com.gym.utils.JwtService;
+import com.gym.login.jwt.JwtService;
 import com.gym.utils.UtilService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
